feat(4615): make editor context fully generalized
